requirements.txt
gradio==5.31.0
koreanize_matplotlib==0.1.1
matplotlib==3.10.3
numpy==2.2.6
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
pandas==2.2.3
Pillow==11.2.1
scikit_learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
timm==1.0.15
tqdm==4.67.1
